sbt-tpolecat (#213)
* sbt-tpolecat

* Attempt to fix e2e tests

* Fix ExchangeInformation model for fapi

+++ src/main/scala/io/github/paoloboni/binance/fapi/response/ExchangeInformation.scala
@@ -59,7 +59,7 @@ case class Symbol(
     baseAssetPrecision: Int,
     quotePrecision: Int,
     underlyingType: String,
-    settlePlan: Int,
+    settlePlan: Long,
     triggerProtect: BigDecimal,
     orderTypes: List[FutureOrderType],
     filters: List[Filter],
